# perf: Add map pool for histogram merges