benchstat agent benches:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                      │ bench-out/agent-ndjson-bench │ bench-out/agent-ndjson-bench-map-pool │
                      │            sec/op            │     sec/op      vs base               │
NDJSON/AgentNodeJS-16                   73.75m ±  3%     48.11m ±  4%  -34.76% (p=0.002 n=6)
NDJSON/AgentPython-16                  103.56m ±  4%     84.29m ± 22%  -18.60% (p=0.026 n=6)
NDJSON/AgentRuby-16                    105.78m ±  2%     69.78m ±  2%  -34.04% (p=0.002 n=6)
NDJSON/AgentGo-16                       181.2m ± 15%     105.7m ±  5%  -41.63% (p=0.002 n=6)
geomean                                 110.0m           73.96m        -32.76%

                      │ bench-out/agent-ndjson-bench │ bench-out/agent-ndjson-bench-map-pool │
                      │             B/op             │      B/op       vs base               │
NDJSON/AgentNodeJS-16                  51.606Mi ± 1%     7.880Mi ± 3%  -84.73% (p=0.002 n=6)
NDJSON/AgentPython-16                   51.85Mi ± 0%     21.34Mi ± 1%  -58.84% (p=0.002 n=6)
NDJSON/AgentRuby-16                     72.69Mi ± 1%     11.95Mi ± 1%  -83.56% (p=0.002 n=6)
NDJSON/AgentGo-16                      129.55Mi ± 2%     15.54Mi ± 4%  -88.01% (p=0.002 n=6)
geomean                                 70.85Mi          13.29Mi       -81.24%

                      │ bench-out/agent-ndjson-bench │ bench-out/agent-ndjson-bench-map-pool │
                      │          allocs/op           │   allocs/op     vs base               │
NDJSON/AgentNodeJS-16                   167.73k ± 0%      94.11k ± 0%  -43.89% (p=0.002 n=6)
NDJSON/AgentPython-16                    287.4k ± 0%      231.6k ± 1%  -19.42% (p=0.002 n=6)
NDJSON/AgentRuby-16                      251.3k ± 1%      147.4k ± 1%  -41.35% (p=0.002 n=6)
NDJSON/AgentGo-16                        326.0k ± 1%      180.6k ± 1%  -44.60% (p=0.002 n=6)
geomean                                  250.7k           155.2k       -38.09%
```

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/histogram-merge-map-pool-before │ bench-out/histogram-merge-map-pool-after │
                            │                  sec/op                   │      sec/op        vs base               │
AggregateCombinedMetrics-16                                 2.119µ ± 6%        2.170µ ± 67%       ~ (p=0.579 n=10)
AggregateBatchSerial-16                                     6.745µ ± 4%        6.942µ ± 45%       ~ (p=0.075 n=10)
AggregateBatchParallel-16                                   7.241µ ± 3%        7.430µ ±  5%  +2.61% (p=0.019 n=10)
CombinedMetricsEncoding-16                                  1.503µ ± 7%        1.520µ ±  8%       ~ (p=0.616 n=10)
CombinedMetricsDecoding-16                                  418.1n ± 2%        418.6n ±  4%       ~ (p=0.566 n=10)
CombinedMetricsToBatch-16                                   271.8µ ± 3%        269.4µ ±  2%       ~ (p=0.579 n=10)
EventToCombinedMetrics-16                                   838.7n ± 4%        848.1n ±  2%       ~ (p=0.393 n=10)
geomean                                                     3.943µ             3.996µ        +1.33%

                            │ bench-out/histogram-merge-map-pool-before │ bench-out/histogram-merge-map-pool-after │
                            │                   B/op                    │      B/op        vs base                 │
AggregateCombinedMetrics-16                              1.531Ki ± 1%        1.490Ki ± 1%  -2.65% (p=0.000 n=10)
AggregateBatchSerial-16                                  2.672Ki ± 3%        2.553Ki ± 4%  -4.46% (p=0.000 n=10)
AggregateBatchParallel-16                                2.671Ki ± 3%        2.512Ki ± 1%  -5.96% (p=0.000 n=10)
CombinedMetricsEncoding-16                                 857.5 ± 3%          864.5 ± 9%       ~ (p=0.447 n=10)
CombinedMetricsDecoding-16                               1.828Ki ± 0%        1.828Ki ± 0%       ~ (p=1.000 n=10) ¹
CombinedMetricsToBatch-16                                37.66Ki ± 0%        37.66Ki ± 0%       ~ (p=1.000 n=10) ¹
EventToCombinedMetrics-16                                  0.000 ± 0%          0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                               ²                    -1.78%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                            │ bench-out/histogram-merge-map-pool-before │ bench-out/histogram-merge-map-pool-after │
                            │                 allocs/op                 │   allocs/op     vs base                  │
AggregateCombinedMetrics-16                                24.00 ± 4%         23.00 ± 4%   -4.17% (p=0.000 n=10)
AggregateBatchSerial-16                                    37.00 ± 0%         31.00 ± 0%  -16.22% (p=0.000 n=10)
AggregateBatchParallel-16                                  37.00 ± 0%         31.00 ± 0%  -16.22% (p=0.000 n=10)
CombinedMetricsEncoding-16                                 0.000 ± 0%         0.000 ± 0%        ~ (p=1.000 n=10) ¹
CombinedMetricsDecoding-16                                 7.000 ± 0%         7.000 ± 0%        ~ (p=1.000 n=10) ¹
CombinedMetricsToBatch-16                                  308.0 ± 0%         308.0 ± 0%        ~ (p=1.000 n=10) ¹
EventToCombinedMetrics-16                                  0.000 ± 0%         0.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean                                                               ²                    -5.51%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```